+++ my-vue-app/src/models/CardsData.js
